@@ -1,70 +1,81 @@
 {
+  "private": true,
+  "license": "MIT",
   "scripts": {
     "build": "gulp build:prod && node createManifest.js && node removeHash.js && yarn svgmin && yarn svgstore",
     "dev": "gulp dev",
     "svgmin": "rm src/img/projects/projects-defs.svg && svgo -f src/img/projects/ --pretty",
     "svgstore": "svgstore -o src/img/projects/projects-defs.svg src/img/projects/*.svg",
-    "svg": "yarn svgmin svgstore"
+    "svg": "yarn svgmin svgstore",
+    "lint": "eslint ."
   },
   "dependencies": {
     "algoliasearch": "^3.21.0",
     "async": "^2.1.4",
-    "babel-core": "^6.4.5",
-    "babel-loader": "^6.2.1",
+    "babel-core": "6.26.0",
+    "babel-eslint": "8.0.2",
+    "babel-loader": "7.1.2",
+    "babel-preset-env": "1.6.1",
     "babel-preset-es2015": "^6.3.13",
     "bootstrap": "4.0.0-alpha.6",
     "browsernizr": "^2.1.0",
     "classie": "^1.0.0",
-    "critical": "^0.8.4",
-    "del": "^2.2.0",
+    "critical": "^1.0.0",
+    "del": "^3.0.0",
     "draggabilly": "^2.1.0",
+    "eslint": "4.11.0",
+    "eslint-config-algolia": "12.0.0",
+    "eslint-config-prettier": "2.8.0",
+    "eslint-plugin-import": "2.8.0",
+    "eslint-plugin-prettier": "2.3.1",
     "favicons": "^4.6.0",
     "glob": "^7.1.2",
     "gulp": "^3.9.1",
-    "gulp-autoprefixer": "^3.1.0",
+    "gulp-autoprefixer": "^4.0.0",
     "gulp-cheerio": "^0.6.2",
     "gulp-critical-css": "^1.0.1",
     "gulp-favicons": "^2.2.5",
     "gulp-filter": "^5.0.0",
     "gulp-gh-pages": "^0.5.4",
-    "gulp-haml": "^0.1.6",
-    "gulp-imagemin": "^2.4.0",
-    "gulp-inject": "^3.0.0",
+    "gulp-haml": "^0.2.0",
+    "gulp-imagemin": "^4.0.0",
+    "gulp-inject": "^4.3.0",
     "gulp-jshint": "^2.0.0",
     "gulp-livereload": "^3.8.1",
     "gulp-minify-css": "^1.2.3",
-    "gulp-notify": "^2.2.0",
+    "gulp-notify": "^3.0.0",
     "gulp-rename": "^1.2.2",
-    "gulp-rev": "^7.1.2",
+    "gulp-rev": "^8.1.0",
     "gulp-rev-napkin": "^0.1.0",
     "gulp-rev-replace": "^0.4.3",
-    "gulp-sass": "^2.1.1",
-    "gulp-scss-lint": "^0.3.9",
+    "gulp-sass": "^3.1.0",
+    "gulp-scss-lint": "^0.5.0",
     "gulp-scss-lint-stylish": "^1.0.0",
     "gulp-svg-sprite": "^1.3.6",
     "gulp-svgmin": "^1.2.0",
-    "gulp-svgstore": "^5.0.5",
+    "gulp-svgstore": "^6.1.0",
     "gulp-sww": "^0.1.4",
-    "gulp-uglify": "^1.5.1",
+    "gulp-uglify": "^3.0.0",
     "gulp-util": "^3.0.8",
     "gulp-webserver": "^0.9.1",
     "html-escape": "^2.0.0",
-    "image-size": "^0.5.4",
-    "imagemin-pngquant": "^4.2.0",
-    "instantsearch.js": "1.11.11",
-    "jquery": "^2.2.0",
+    "image-size": "^0.6.1",
+    "imagemin-pngquant": "5.0.1",
+    "instantsearch.js": "2.2.5",
+    "jquery": "^3.2.1",
     "jshint": "^2.9.1",
     "jshint-stylish": "^2.1.0",
     "json-loader": "^0.5.4",
     "lodash": "^4.17.4",
     "mo-js": "^0.288.1",
-    "run-sequence": "^1.1.5",
-    "sw-toolbox": "^3.6.0",
-    "webpack": "^1.12.9",
-    "webpack-stream": "^3.1.0",
+    "prettier": "1.8.2",
+    "run-sequence": "^2.2.0",
     "svgmin": "0.1.0",
     "svgo": "1.0.3",
-    "svgstore": "2.0.2"
+    "svgstore": "2.0.2",
+    "sw-toolbox": "^3.6.0",
+    "webpack": "3.8.1",
+    "webpack-stream": "^4.0.0"
   },
   "engines": {
     "node": "^8.9.1"
