@@ -1,67 +1,70 @@
-
 const templates = require('./templates.js');
-const instantsearch = require('instantsearch.js');
+import instantsearch from 'instantsearch.js';
+import { searchBox, menu, hits } from 'instantsearch.js/es/widgets';
 
-const projects = require('./../algolia-projects.json');
+const algoliaProjects = require('./../algolia-projects.json');
 const config = require('./../../config.json');
 
 let { appID, apiKey, index } = config.algolia;
 
-appID = "HXQH62TCI4";
-apiKey = "0b9f3069b37517348a864b7239a8abfa";
-index = "community";
+appID = 'HXQH62TCI4';
+apiKey = '0b9f3069b37517348a864b7239a8abfa';
+index = 'community';
 
 const loadDefs = () => {
   fetch('/img/projects/projects-defs.svg')
     .then(r => r.text())
     .then(svg => {
       document.querySelector('.svg-icons').innerHTML = svg;
-    })
-}
+    });
+};
 
 function validateEmail(email) {
-  var re = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
-  return re.test(email)
+  /* eslint-disable */
+  const re = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+  return re.test(email);
 }
 
-window.addEventListener('load',() => {
-
+window.addEventListener('load', () => {
   loadDefs();
 
-  if(window.location.hash) {
+  if (window.location.hash) {
     const hash = document.querySelector(`[href="${window.location.hash}"]`);
     setTimeout(() => {
       scrollToElement(hash, 400);
-    }, 60)
+    }, 60);
   }
 });
 
-const easeInOutCubic = (t, b, c, d) => {
-  return -c / 2 * (Math.cos(Math.PI * t / d) - 1) + b
-}
+const easeInOutCubic = (t, b, c, d) =>
+  -c / 2 * (Math.cos(Math.PI * t / d) - 1) + b;
 
 const parentWithClass = (startingNode, matchClassName) => {
   let node = startingNode;
-  while(node.parentNode){
-    if(node.className.split(' ').indexOf(matchClassName) > -1){
+  while (node.parentNode) {
+    if (node.className.split(' ').indexOf(matchClassName) > -1) {
       return node;
     }
     node = node.parentNode;
   }
-}
+};
 
-function scrollToElement(nodeOrSelector, totalTime = 400){
-  let node,
-      scrollToNode,
-      time = {};
+function scrollToElement(nodeOrSelector, totalTime = 400) {
+  let node;
+  let scrollToNode;
+  const time = {};
 
   time.start = performance.now();
   time.total = totalTime;
 
-  if(typeof nodeOrSelector === "string"){
+  if (typeof nodeOrSelector === 'string') {
     node = document.querySelectorAll(nodeOrSelector);
-    if(node.length > 1){
-      throw new Error(`Selector matches ${node.length} elements, please provide a unique selector`);
+    if (node.length > 1) {
+      throw new Error(
+        `Selector matches ${
+          node.length
+        } elements, please provide a unique selector`
+      );
     }
     scrollToNode = node[0];
   } else {
@@ -73,63 +76,74 @@ function scrollToElement(nodeOrSelector, totalTime = 400){
 
   const tick = now => {
     const elapsed = now - time.start;
-    const position = Math.round(easeInOutCubic(elapsed, startDistance, endDistance, time.total));
+    const position = Math.round(
+      easeInOutCubic(elapsed, startDistance, endDistance, time.total)
+    );
     window.scroll(0, position);
-    elapsed < time.total ? window.requestAnimationFrame(tick) : null;
-  }
+    if (elapsed < time.total) {
+      window.requestAnimationFrame(tick);
+    }
+  };
 
   window.requestAnimationFrame(tick);
 }
 
-const sortProjectsByCategory = (projects) => {
-  let sorted = {};
+const sortProjectsByCategory = projects => {
+  const sorted = {};
   projects.sort((a, b) => {
-    const catA = a.category,
-          catB = b.category;
-    if(catA > catB) return -1;
-    else if ( catA < catB ) return 1;
+    const catA = a.category;
+    const catB = b.category;
+    if (catA > catB) return -1;
+    else if (catA < catB) return 1;
     return 0;
   });
 
-  projects.forEach((p, index) => {
-    if(!sorted[p.category]){
+  projects.forEach(p => {
+    if (!sorted[p.category]) {
       sorted[p.category] = [];
     }
     sorted[p.category].push(p);
   });
   return sorted;
-}
+};
 
-const renderResults = (projects) => {
-  const injectInside = document.querySelector(".alg-communityprojects__hits");
-  let sortedProjects = [];
+const renderItem = item => {
+  const wrapperDiv = document.createElement('div');
+  wrapperDiv.className = 'ais-hits--item';
+  wrapperDiv.innerHTML = templates.hitTemplate(item);
+  return wrapperDiv;
+};
+
+const renderResults = projects => {
+  const injectInside = document.querySelector('.alg-communityprojects__hits');
+  const sortedProjects = [];
 
   Object.keys(projects).forEach(cat => {
-    sortedProjects.push(projects[cat])
+    sortedProjects.push(projects[cat]);
   });
 
-  sortedProjects.sort((a,b) => a.length < b.length);
+  sortedProjects.sort((a, b) => a.length < b.length);
 
   sortedProjects.forEach(p => {
-    p.sort((a,b) => a.ranking < b.ranking);
-  })
+    p.sort((a, b) => a.ranking < b.ranking);
+  });
 
   sortedProjects.forEach(projectsArray => {
-    let cat = projectsArray[0].category || "Misc"
-    let categoryArray = projects[cat]
-    let dummyArray = projects[cat]
-    let viewMore = null
+    const cat = projectsArray[0].category || 'Misc';
+    const categoryArray = projects[cat];
+    const dummyArray = projects[cat];
+    let viewMore = null;
 
-    if(categoryArray.length > 4){
+    if (categoryArray.length > 4) {
       viewMore = dummyArray.length - 4;
     }
 
-    const wrapper = document.createElement("div");
-    wrapper.className = "alg-communityprojects__hitswrapper";
+    const wrapper = document.createElement('div');
+    wrapper.className = 'alg-communityprojects__hitswrapper';
     wrapper.innerHTML = templates.headerTemplate(cat, viewMore);
     const wrapperHits = wrapper.querySelector('.ais-hits');
 
-    if(categoryArray && categoryArray.length) {
+    if (categoryArray && categoryArray.length) {
       categoryArray.forEach(project => {
         const article = renderItem(project);
         wrapperHits.appendChild(article);
@@ -137,47 +151,10 @@ const renderResults = (projects) => {
     }
     injectInside.appendChild(wrapper);
   });
-}
-
-const renderItem = (data) => {
-  const wrapperDiv = document.createElement("div");
-  wrapperDiv.className = "ais-hits--item";
-  wrapperDiv.innerHTML = templates.hitTemplate(data);
-  return wrapperDiv;
-}
-
-const renderMenuList = (projects) => {
-  const listContainer = document.querySelector(".alg-communityprojects__facets ul");
-  const totalProjects = Object.keys(projects).reduce((prev, current) => prev += projects[current].length, 0);
-  const totalLi = renderMenuItem("All Projects", totalProjects, true);
-  listContainer.appendChild(totalLi);
-
-  let sortedProjects = [];
-  Object.keys(projects).forEach(type => {
-    const typeCategory = projects[type];
-    sortedProjects.push(typeCategory);
-  });
-
-  sortedProjects.sort((a,b) => a.length < b.length);
-
-  sortedProjects.forEach(projectArray => {
-    if(projectArray && projectArray.length){
-      const li = renderMenuItem(projectArray[0].category || "Misc", projectArray.length)
-      listContainer.appendChild(li);
-    }
-  })
-}
-
-const renderMenuItem = (category, count, isHeader) => {
-  const li = document.createElement("li");
-  li.className = "ais-menu--item";
-  li.innerHTML = templates.menuTemplate(category, count, isHeader)
-  return li;
-}
-
+};
 
 window.pardotAppendIframe = function pardotAppendIframe(url) {
-  var iframe = document.createElement('iframe');
+  const iframe = document.createElement('iframe');
   iframe.src = url;
   iframe.width = 1;
   iframe.height = 1;
@@ -185,145 +162,143 @@ window.pardotAppendIframe = function pardotAppendIframe(url) {
 };
 
 function submitPardotForm(event) {
-  event.preventDefault()
-  var $signupForm = this.parentNode
-  var $message = this.querySelector('.mc-signupmessage')
-  var size = $signupForm.querySelector('input').getBoundingClientRect()
-  var email = this.querySelector('#mce-EMAIL').value
-  var action =
+  event.preventDefault();
+  const $signupForm = this.parentNode;
+  const $message = this.querySelector('.mc-signupmessage');
+  const size = $signupForm.querySelector('input').getBoundingClientRect();
+  const email = this.querySelector('#mce-EMAIL').value;
+  const action =
     this.getAttribute('action') ||
-    'https://go.pardot.com/l/139121/2016-06-09/f3kzm'
+    'https://go.pardot.com/l/139121/2016-06-09/f3kzm';
 
-  const isValidEmail = validateEmail(email)
+  const isValidEmail = validateEmail(email);
 
-  if (!isValidEmail) return false
+  if (!isValidEmail) return false;
 
   if (email) {
-    pardotAppendIframe(action + '?email=' + encodeURI(email))
-    $signupForm.classList.toggle('newsletter-signup--success')
+    window.pardotAppendIframe(`${action}?email=${encodeURI(email)}`);
+    $signupForm.classList.toggle('newsletter-signup--success');
     $message.innerHTML =
-      'Thank you for subscribing <svg xmlns="http://www.w3.org/2000/svg" width="36" height="35" viewBox="0 0 36 35"><path d="M18 4.02C16.552 2.25 14.057.297 10.103.297 3.373.297 0 6.63 0 12.547c0 9.776 15.135 20.15 16.99 21.435.607.422 1.413.422 2.02 0C20.864 32.695 36 22.322 36 12.546 36 6.63 32.627.298 25.896.298c-3.953 0-6.448 1.954-7.896 3.72v.002z" fill="#fb366e" fill-rule="evenodd"/></svg>'
-    $message.style.width = size.width + 'px'
-    $message.style.height = size.height + 'px'
+      'Thank you for subscribing <svg xmlns="http://www.w3.org/2000/svg" width="36" height="35" viewBox="0 0 36 35"><path d="M18 4.02C16.552 2.25 14.057.297 10.103.297 3.373.297 0 6.63 0 12.547c0 9.776 15.135 20.15 16.99 21.435.607.422 1.413.422 2.02 0C20.864 32.695 36 22.322 36 12.546 36 6.63 32.627.298 25.896.298c-3.953 0-6.448 1.954-7.896 3.72v.002z" fill="#fb366e" fill-rule="evenodd"/></svg>';
+    $message.style.width = `${size.width}px`;
+    $message.style.height = `${size.height}px`;
   }
+
+  return true;
 }
 
-const subscribeForm = document.querySelector('#mc-embedded-subscribe-form')
-subscribeForm.addEventListener('submit', submitPardotForm)
+const subscribeForm = document.querySelector('#mc-embedded-subscribe-form');
+subscribeForm.addEventListener('submit', submitPardotForm);
 
-const addTagToHelper = (event) => {
-  search.helper.toggleRefine('category', event.target.dataset.tag);
-  search.helper.search();
-}
-
-const DefaultContainer = document.querySelector('.alg-communityprojects__hits.default');
-const SearchContainer = document.querySelector('.alg-communityprojects__hits.is');
+const DefaultContainer = document.querySelector(
+  '.alg-communityprojects__hits.default'
+);
+const SearchContainer = document.querySelector(
+  '.alg-communityprojects__hits.is'
+);
 const ClearRefinements = document.querySelector('.alg-community--clearSearch');
 
-let search = instantsearch({
+const search = instantsearch({
   appId: appID,
-  apiKey: apiKey,
+  apiKey,
   indexName: index,
-  searchFunction: (helper) => {
-    const hasCategoryRefinement = helper.state.hierarchicalFacetsRefinements.category && helper.state.hierarchicalFacetsRefinements.category.length > 0;
-    if(helper.state.query == ""){
+  searchFunction: helper => {
+    const hasCategoryRefinement =
+      helper.state.hierarchicalFacetsRefinements.category &&
+      helper.state.hierarchicalFacetsRefinements.category.length > 0;
+    if (helper.state.query == '') {
       ClearRefinements.classList.remove('visible');
     } else {
       ClearRefinements.classList.add('visible');
     }
-    if(helper.state.query == "" && !hasCategoryRefinement) {
-      DefaultContainer.style.display = "block";
-      SearchContainer.style.display = "none";
+    if (helper.state.query == '' && !hasCategoryRefinement) {
+      DefaultContainer.style.display = 'block';
+      SearchContainer.style.display = 'none';
     } else {
-      DefaultContainer.style.display = "none";
-      SearchContainer.style.display = "block";
+      DefaultContainer.style.display = 'none';
+      SearchContainer.style.display = 'block';
     }
-    helper.search()
-  }
+    helper.search();
+  },
 });
 
-// search.addWidget(
-//   instantsearch.widgets.clearAll()
-// )
-
 search.addWidget(
-  instantsearch.widgets.searchBox({
+  searchBox({
     container: '#alg-community__search',
     placeholder: 'Search for projects, libraries, plugins, demos...',
-    autofocus: false
+    autofocus: false,
   })
 );
 
 search.addWidget(
-  instantsearch.widgets.menu({
+  menu({
     container: '.alg-communityprojects__facets__is',
     attributeName: 'category',
     limit: 10,
     templates: {
-      item: templates.menuTemplate_is,
-      header: templates.header(projects.length)
-    }
+      item: templates.menuTemplateIs,
+      header: templates.header(algoliaProjects.length),
+    },
   })
 );
 
 // add a hits widget
 search.addWidget(
-  instantsearch.widgets.hits({
+  hits({
     container: '.alg-communityprojects__hits.is',
     hitsPerPage: 30,
-    templates:{
+    templates: {
       item: templates.hitTemplate,
-      empty: templates.noHits
-    }
+      empty: templates.noHits,
+    },
   })
 );
 
 function onViewMoreClick(event) {
   event.preventDefault();
-  const parentWrapper = parentWithClass(this, "alg-communityprojects__hitswrapper");
-  parentWrapper.classList.contains('expanded') ?
-  parentWrapper.classList.remove('expanded') :
-  parentWrapper.classList.add('expanded');
+  const parentWrapper = parentWithClass(
+    this,
+    'alg-communityprojects__hitswrapper'
+  );
+  if (parentWrapper.classList.contains('expanded')) {
+    parentWrapper.classList.remove('expanded');
+  } else {
+    parentWrapper.classList.add('expanded');
+  }
   scrollToElement(this);
-};
+}
 
-const sorted = sortProjectsByCategory(projects);
+const sorted = sortProjectsByCategory(algoliaProjects);
 renderResults(sorted);
 
 const viewMoreLinks = [...document.querySelectorAll('.alg-viewmore')];
 viewMoreLinks.forEach(l => l.addEventListener('click', onViewMoreClick));
 
 search.on('render', () => {
-
   const header = document.querySelector('[data-tag="All Projects"]');
 
-  header.addEventListener('click', (e) => {
+  header.addEventListener('click', e => {
     e.preventDefault();
     search.helper.clearRefinements();
     search.helper.setQuery('');
     search.helper.search();
   });
 
-  ClearRefinements.addEventListener('click', (e) => {
+  ClearRefinements.addEventListener('click', e => {
     e.preventDefault();
     search.helper.clearRefinements();
     search.helper.setQuery('');
     search.helper.search();
   });
-
-})
+});
 
 search.start();
 
 if ('serviceWorker' in navigator) {
   navigator.serviceWorker
     .register('/serviceWorker.js', { scope: '/' })
-    .then(function(registration) {
-    })
-    .catch(function(error) {
-      console.log(error)
-    })
+    .then(() => {})
+    .catch(error => {
+      console.log(error); // eslint-disable-line no-console
+    });
 }
-
-
-
