@@ -1,41 +1,28 @@
 const gulp = require('gulp');
-const path = require('path');
 const del = require('del');
-const rename = require("gulp-rename");
 const runSequence = require('run-sequence');
 const glob = require('glob');
 
-//dev
+// dev
 const webserver = require('gulp-webserver');
 const livereload = require('gulp-livereload');
 
-//deploy
+// deploy
 const ghPages = require('gulp-gh-pages');
 
-//css
+// css
 const sass = require('gulp-sass');
-const scsslint = require('gulp-scss-lint');
-const scssLintStylish = require('gulp-scss-lint-stylish');
 const autoprefixer = require('gulp-autoprefixer');
 const critical = require('critical').stream;
 
-//html
+// html
 const haml = require('gulp-haml');
 
-//serviceworker
-// const sww = require('gulp-sww');
-
-//SVG
-const svgstore = require('gulp-svgstore');
-const svgmin = require('gulp-svgmin');
-const cheerio = require('gulp-cheerio');
-
-//js
+// js
 const webpack = require('webpack-stream');
-const webpackConfig = require("./webpack.config.js");
-const jshint = require('gulp-jshint');
+const webpackConfig = require('./webpack.config.js');
 
-//prod
+// prod
 const rev = require('gulp-rev');
 const revNapkin = require('gulp-rev-napkin');
 const revReplace = require('gulp-rev-replace');
@@ -43,10 +30,10 @@ const revReplace = require('gulp-rev-replace');
 const minifyCss = require('gulp-minify-css');
 const uglify = require('gulp-uglify');
 const imagemin = require('gulp-imagemin');
-const pngquant = require('imagemin-pngquant')
+const pngquant = require('imagemin-pngquant');
 
-//debug utils
-var gutil = require('gulp-util');
+// debug utils
+const gutil = require('gulp-util');
 
 const algolia = require('algoliasearch');
 const config = require('./config.json');
@@ -63,107 +50,95 @@ const fs = require('fs');
 //
 // *************************************
 
-
 // -------------------------------------
 //   Task: Clean build directory
 // -------------------------------------
-gulp.task('clean', function() {
-  return del("build");
-});
+gulp.task('clean', () => del('build'));
 
 // -------------------------------------
 //   Task: Haml
 // -------------------------------------
-gulp.task('haml', function() {
-  return gulp.src('src/*.haml')
+gulp.task('haml', () =>
+  gulp
+    .src('src/*.haml')
     .pipe(haml())
-    .pipe(gulp.dest('build'));
-});
-
+    .pipe(gulp.dest('build'))
+);
 
 // -------------------------------------
 //   Task: SCSS
 // -------------------------------------
-gulp.task('scss', function() {
-  return gulp.src('src/scss/**/*.scss')
+gulp.task('scss', () =>
+  gulp
+    .src('src/scss/**/*.scss')
     .pipe(sass().on('error', sass.logError))
-    .pipe(autoprefixer({
-      browsers: ['last 2 versions'],
-      cascade: false
-    }))
+    .pipe(
+      autoprefixer({
+        browsers: ['last 2 versions'],
+        cascade: false,
+      })
+    )
     .pipe(gulp.dest('build'))
-    .pipe(livereload());
-});
+    .pipe(livereload())
+);
 
-gulp.task('critical-css', function() {
+gulp.task('critical-css', () => {
   const cssFiles = glob.sync('build/*.css');
   const htmlFiles = glob.sync('build/*.html');
 
-  return gulp.src(htmlFiles)
-    .pipe(critical({ 
-      base: 'build/',
-      inline: true, 
-      minify: true,
-      css: [ ...cssFiles ],
-      extract: true
-    }))
-    .on('error', function(err) { gutil.log(gutil.colors.red(err.message)); })
+  return gulp
+    .src(htmlFiles)
+    .pipe(
+      critical({
+        base: 'build/',
+        inline: true,
+        minify: true,
+        css: [...cssFiles],
+        extract: true,
+      })
+    )
+    .on('error', err => {
+      gutil.log(gutil.colors.red(err.message));
+    })
     .pipe(gulp.dest('build'));
 });
-
-// -------------------------------------
-//   Task: Lint SCSS
-// -------------------------------------
-// gulp.task('lint:scss', function() {
-//   return gulp.src('src/scss/*.scss')
-//     .pipe(scsslint({
-//       customReport: scssLintStylish,
-//       config: 'scss-lint.yml'
-//     }));
-// });
 
 // -------------------------------------
 //   Task: Minify CSS
 // -------------------------------------
-gulp.task('css:min', ['scss'], function() {
-  return gulp.src('build/*.css')
+gulp.task('css:min', ['scss'], () =>
+  gulp
+    .src('build/*.css')
     .pipe(minifyCss())
-    .pipe(gulp.dest('build'));
-});
+    .pipe(gulp.dest('build'))
+);
 
 // -------------------------------------
 //   Task: Javascript
 // -------------------------------------
-gulp.task('js', function() {
-  return gulp.src(['src/js/app.js','src/serviceWorker.js'])
+gulp.task('js', () =>
+  gulp
+    .src(['src/js/app.js', 'src/serviceWorker.js'])
     .pipe(webpack(webpackConfig))
-    .on('error', function(e) {
+    .on('error', function() {
       this.emit('end'); // Recover from errors
     })
     .pipe(gulp.dest('build/js'))
-    .pipe(livereload());
-});
-
-// -------------------------------------
-//   Task: Lint Javascript
-// -------------------------------------
-gulp.task('lint:js', function() {
-  return gulp.src('src/js/*.js')
-    .pipe(jshint())
-    .pipe(jshint.reporter('jshint-stylish'));
-});
+    .pipe(livereload())
+);
 
 // -------------------------------------
 //   Task: Minify JS
 // -------------------------------------
-gulp.task('js:min', ['js'], function() {
-  return gulp.src('build/js/*.js')
+gulp.task('js:min', ['js'], () =>
+  gulp
+    .src('build/js/*.js')
     .pipe(uglify())
-    .pipe(gulp.dest('build/js'));
-});
+    .pipe(gulp.dest('build/js'))
+);
 
 // -------------------------------------
-//   Task: Initialise ServiceWorker 
+//   Task: Initialise ServiceWorker
 // -------------------------------------
 // gulp.task('offline', function() {
 //   return gulp.src('**/*', { cwd: 'build' })
@@ -174,106 +149,142 @@ gulp.task('js:min', ['js'], function() {
 // -------------------------------------
 //   Task: Images & Optmizations
 // -------------------------------------
-gulp.task('images', function() {
-  return gulp.src('src/img/**/*')
+gulp.task('images', () =>
+  gulp
+    .src('src/img/**/*')
     .pipe(gulp.dest('build/img'))
-    .pipe(livereload());
-});
+    .pipe(livereload())
+);
 
-gulp.task('images:optim', function() {
-  return gulp.src('src/img/**/*')
-    .pipe(imagemin({
-      progressive: true,
-      use: [pngquant()]
-    }))
-    .pipe(gulp.dest('build/img'));
-});
+gulp.task('images:optim', () =>
+  gulp
+    .src('src/img/**/*')
+    .pipe(
+      imagemin({
+        progressive: true,
+        use: [pngquant()],
+      })
+    )
+    .pipe(gulp.dest('build/img'))
+);
 
 // -------------------------------------
 //   Task: Copy files
 // -------------------------------------
-gulp.task('copy', function() {
-  return gulp.src(['CNAME'])
-    .pipe(gulp.dest('build'));
-});
+gulp.task('copy', () => gulp.src(['CNAME']).pipe(gulp.dest('build')));
 
-gulp.task('copyRedirects', function() {
-  return gulp.src(['_redirects'])
-    .pipe(gulp.dest('build'));
-});
+gulp.task('copyRedirects', () =>
+  gulp.src(['_redirects']).pipe(gulp.dest('build'))
+);
 
-gulp.task('copyWorker', function() {
-  return gulp.src(['build/js/serviceWorker.js'])
-    .pipe(gulp.dest('build'));
-});
+gulp.task('copyWorker', () =>
+  gulp.src(['build/js/serviceWorker.js']).pipe(gulp.dest('build'))
+);
 
 // -------------------------------------
 //   Task: Watch
 // -------------------------------------
-gulp.task('watch', function() {
+gulp.task('watch', () => {
   livereload.listen();
   gulp.watch('src/*.haml', ['build:haml']);
   gulp.watch('src/scss/**/*.scss', ['scss']);
-  gulp.watch('src/js/**/*.js', ['js', 'lint:js']);
+  gulp.watch('src/js/**/*.js', ['js']);
   gulp.watch('src/img/**/*', ['images']);
 });
 
 // -------------------------------------
 //   Task: Web Server
 // -------------------------------------
-gulp.task('webserver', function() {
-  gulp.src('build')
-    .pipe(webserver({
+gulp.task('webserver', () => {
+  gulp.src('build').pipe(
+    webserver({
       host: 'localhost',
       port: 1337,
       livereload: true,
       directoryListing: false,
-      open: true
-    }));
+      open: true,
+    })
+  );
 });
 
 // -------------------------------------
 //   Task: Revision
 // -------------------------------------
-gulp.task('revision', function() {
-  return gulp.src(['build/**/**/*.{css,js,png,svg,html}','!build/js/serviceWorker.js', '!build/img/*.svg'])
+gulp.task('revision', () =>
+  gulp
+    .src([
+      'build/**/**/*.{css,js,png,svg,html}',
+      '!build/js/serviceWorker.js',
+      '!build/img/*.svg',
+    ])
     .pipe(rev())
     .pipe(gulp.dest('build'))
     .pipe(revNapkin())
     .pipe(rev.manifest())
     .pipe(gulp.dest('build'))
-});
+);
 
 gulp.task('revreplace', ['revision'], () => {
-  const manifest = gulp.src("build/rev-manifest.json");
-  const htmlFiles = glob.sync('build/**/*')
+  const manifest = gulp.src('build/rev-manifest.json');
 
-  return gulp.src(['build/*.html', 'build/js/*.js','build/*.js', 'build/*.css', 'build/manifest.json'],{ base: 'build' })
-    .pipe(revReplace({ 
-      manifest: manifest 
-    }))
-    .pipe(gulp.dest('build'))
-})
+  return gulp
+    .src(
+      [
+        'build/*.html',
+        'build/js/*.js',
+        'build/*.js',
+        'build/*.css',
+        'build/manifest.json',
+      ],
+      { base: 'build' }
+    )
+    .pipe(
+      revReplace({
+        manifest,
+      })
+    )
+    .pipe(gulp.dest('build'));
+});
 
 // -------------------------------------
 //   Task: Build DEV - PROD - HAML
 // -------------------------------------
-gulp.task('build:dev', ['clean'], function(callback) {
-  runSequence('scss', 'images', 'haml', 'js', 'copyRedirects', 'copyWorker', callback);
+gulp.task('build:dev', ['clean'], callback => {
+  runSequence(
+    'scss',
+    'images',
+    'haml',
+    'js',
+    'copyRedirects',
+    'copyWorker',
+    callback
+  );
 });
 
-gulp.task('build:prod', ['clean'], function(callback) {
-  runSequence('scss', 'css:min', 'copy', 'images:optim', 'haml', 'js:min', 'revreplace', 'critical-css', 'copyRedirects', 'copyWorker', callback);
+gulp.task('build:prod', ['clean'], callback => {
+  runSequence(
+    'scss',
+    'css:min',
+    'copy',
+    'images:optim',
+    'haml',
+    'js:min',
+    'revreplace',
+    'critical-css',
+    'copyRedirects',
+    'copyWorker',
+    callback
+  );
 });
 
-gulp.task('build:haml', function(callback) {
+gulp.task('build:haml', callback => {
   runSequence('haml', callback);
 });
 
 // -------------------------------------
 //   Task: Developement
 // -------------------------------------
-gulp.task('dev', function(callback) {
+gulp.task('dev', callback => {
   runSequence('build:dev', 'watch', 'webserver', callback);
 });
 
@@ -281,41 +292,57 @@ gulp.task('dev', function(callback) {
 //   Task: Algolia
 // -------------------------------------
 
-gulp.task('export:algolia-index', function() {
-  var client = algolia(config.algolia.appID, config.algolia.apiKey);
-  var index = client.initIndex(config.algolia.index);
-  fs.readFile('src/algolia-projects.json', 'utf8', function(err, data) {
+gulp.task('export:algolia-index', () => {
+  /* eslint-disable */
+  const client = algolia(config.algolia.appID, config.algolia.apiKey);
+  const index = client.initIndex(config.algolia.index);
+  fs.readFile('src/algolia-projects.json', 'utf8', (err, data) => {
     if (err) {
-      return console.log(err);
+      return console.log(err); // eslint-disable-line no-console
     }
-    var data = _.chunk(JSON.parse(data), [size = 100]);
-    return index.clearIndex(function(err, content) {
-      index.waitTask(content.taskID, function() {
-        async.each(data, function(batch, callback) {
-          index.addObjects(batch, function(err, result) {
-            index.waitTask(result.taskID, function() {
-              console.log('Indexed ' + batch.length + ' objects');
-              callback();
+    const parsedData = _.chunk(JSON.parse(data), [(size = 100)]);
+    return index.clearIndex((err, content) => {
+      if (err) {
+        return console.log(err); // eslint-disable-line no-console
+      }
+      index.waitTask(content.taskID, err => {
+        if (err) {
+          return console.log(err); // eslint-disable-line no-console
+        }
+        async.each(
+          parsedData,
+          (batch, callback) => {
+            index.addObjects(batch, (err, result) => {
+              if (err) {
+                return console.log(err); // eslint-disable-line no-console
+              }
+              index.waitTask(result.taskID, err => {
+                if (err) {
+                  return console.log(err); // eslint-disable-line no-console
+                }
+                console.log(`Indexed ${batch.length} objects`); // eslint-disable-line no-console
+                callback();
+              });
             });
-          });
-        }, function(err) {
-          console.log(err);
-        });
+          },
+          err => {
+            console.log(err); // eslint-disable-line no-console
+          }
+        );
       });
     });
   });
 });
 
-gulp.task('export:algolia-settings', function() {
-  var client = algolia(config.algolia.appID, config.algolia.apiKey);
-  var index = client.initIndex(config.algolia.index);
-  index.setSettings(config.algolia.settings, function(err, content) {});
+gulp.task('export:algolia-settings', () => {
+  const client = algolia(config.algolia.appID, config.algolia.apiKey);
+  const index = client.initIndex(config.algolia.index);
+  index.setSettings(config.algolia.settings, (err, content) => {});
 });
 
 // -------------------------------------
 //   Task: Deploy Github Page
 // -------------------------------------
-gulp.task('deploy', function(callback) {
-  return gulp.src('build/**/*')
-    .pipe(ghPages({ branch: "master" }));
-});
+gulp.task('deploy', callback =>
+  gulp.src('build/**/*').pipe(ghPages({ branch: 'master' }))
+);
