@@ -1,34 +1,35 @@
-
-var webpack = require("webpack");
-var path = require("path");
+const webpack = require('webpack');
 
 module.exports = {
-    entry: {
-        app: './src/js/app.js',
-        serviceWorker: './src/js/serviceWorker.js'
-    },
-    output: {
-        path: __dirname,
-        filename: '[name].js'
-    },
-    module: {
-        loaders: [
-            {
-              test: /\.jsx?$/,
-              exclude: /(node_modules|bower_components)/,
-              loader: 'babel-loader?presets[]=es2015'
-            },
-            {
-                test: /\.json$/,
-                loader: 'json-loader'
-            }
-        ]
-    },
-    plugins: [
-        new webpack.DefinePlugin({
-            "process.env": {
-                'NODE_ENV': '"production"'
-            }
-        })
-    ]
+  entry: {
+    app: './src/js/app.js',
+    serviceWorker: './src/js/serviceWorker.js',
+  },
+  output: {
+    path: __dirname,
+    filename: '[name].js',
+  },
+  module: {
+    rules: [
+      {
+        test: /\.jsx$/,
+        exclude: /(node_modules|bower_components)/,
+        use: {
+          loader: 'babel-loader',
+          options: {
+            presets: ['@babel/preset-env'],
+          },
+        },
+      },
+      {
+        test: /\.json$/,
+        loader: 'json-loader',
+      },
+    ],
+  },
+  plugins: [
+    new webpack.EnvironmentPlugin({
+      NODE_ENV: 'production',
+    }),
+  ],
 };
